Whole bunch of ASD

+++ Smt 2/Algoritma dan Struktur Data/Praktikum/week 5/Faktorial/Faktorial.java
@@ -1,22 +1,23 @@
 package Faktorial;
 
 public class Faktorial {
-    public int num;
-    public int faktorialBF(int n) {
-        int fakto = 1;
-        for (int i = 1; i <= n; i++) {
+    public long num;
+    public long faktorialBF(long n) {
+        long fakto = 1;
+        for (long i = 1; i <= n; i++) {
             fakto = fakto * i;
         }
         return fakto;
     }
-    public int faktorialDC(int n) {
+    public long faktorialDC(long n) {
         if (n==1) {
             return 1;
         }
         else
         {
-            int fakto = n * faktorialDC(n-1);
+            long fakto = n * faktorialDC(n-1);
             return fakto;
         }
     }
+    
 }
